+++ rod/libs/py/tf/tfvars.py
@@ -103,6 +103,7 @@ class Cloud(BaseTfVarsModel):
 class Env(BaseTfVarsModel):
     name: str
     short_name: str
+    type: Literal["internal", "product"]
     initial_start: bool = False
     users: dict[str, User]
     apps: dict[str, App]
@@ -136,6 +137,19 @@ class TfVars(BaseTfVarsModel):
     repo: Repo
     ci: Ci
     envs: dict[str, Env]
+
+    @model_validator(mode="after")
+    def validate_single_internal_env(self):
+        internal_envs = [
+            env_name
+            for env_name, env_obj in self.envs.items()
+            if env_obj.type == "internal"
+        ]
+
+        if len(internal_envs) != 1:
+            raise ValueError('exactly one env type must be "internal"')
+
+        return self
 
 
 def tfvars():